requirements.txt
peft==0.7.1
transformers==4.36.0
datasets==2.14.5
torch==2.6.0
pandas==2.1.1
numpy==1.26.0
tqdm==4.66.1
requests==2.31.0
ollama-python==0.1.2
pytz==2023.3
astral==3.2
